+++ src/test/java/fibonacci.java
@@ -1,19 +1,15 @@
-import java.io.IOException;
+import java.io.File;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.util.Iterator;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.parsers.ParserConfigurationException;
 
-import org.jaxen.JaxenException;
 import org.jaxen.XPath;
 import org.jaxen.dom.DOMXPath;
 import org.jaxen.dom.DocumentNavigator;
 import org.w3c.dom.Node;
 import org.xml.sax.InputSource;
-import org.xml.sax.SAXException;
 
 import be.baur.sdt.SDT;
 import be.baur.sdt.TransformContext;
@@ -22,8 +18,7 @@ import be.baur.sdt.transform.Transform;
 @SuppressWarnings("rawtypes")
 public class fibonacci {
 
-	public static void main(String[] args) 
-		throws ParserConfigurationException, SAXException, IOException, JaxenException {
+	public static void main(String[] args) throws Exception {
 		
 		// with W3C DOM
 		DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();   
@@ -45,15 +40,15 @@ public class fibonacci {
 		}
 
 		// with SDT
-		InputStream in = transform.class.getResourceAsStream("/fibonacci.sdt");
-		Transform tran = SDT.parse(new InputStreamReader(in, "UTF-8"));
+		String sdtfile = transform.class.getResource("/fibonacci.sdt").getFile();
+		Transform tran = SDT.parse(new File(sdtfile));
 		
-		String file = transform.class.getResource("/fibonacci.xml").getFile();
+		String sdafile = transform.class.getResource("/fibonacci.xml").getFile();
 		TransformContext context = new TransformContext.Builder()
 			.addNamespace("f", "http://namespaces.cafeconleche.org/xmljava/ch3/")
 			.addNamespace("SOAP", "http://schemas.xmlsoap.org/soap/envelope/")
 			.setNavigator(DocumentNavigator.getInstance())
-			.setStringParameter("filename", file).build();
+			.setStringParameter("filename", sdafile).build();
 		tran.execute(context); 
 		
 	}