+++ RELEASE_NOTES.md
@@ -5,6 +5,9 @@
 - `Removed` SDT namespace prefix and URI declarations from the SDT class.
 - `Renamed` sdt.xpath.functions package to sdt.functions.
 - `Renamed` StatementContext.setContextNode() to setXPathContext().
+- Added TransformContext.Builder.setNavigator().
+- Added SDTNamespaceContext.addNamespace().
+
 
 ## Compatibility